+++ Profitable/Profitable.Services/Posts/PostService.cs
@@ -35,7 +35,7 @@ namespace Profitable.Services.Posts
         }
 
         public async Task<Result> AddPostAsync(
-            ApplicationUser author,
+            Guid authorId,
             AddPostRequestModel newPost)
         {
             if (newPost.Content.Length > GlobalServicesConstants.PostMaxLength)
@@ -50,7 +50,8 @@ namespace Profitable.Services.Posts
                 var postToAdd = mapper.Map<Post>(newPost, opt => 
                 opt.AfterMap((src, dest) => 
                 {
-                    dest.Author = author;
+                    dest.AuthorId = authorId;
+                    dest.PostedOn = DateTime.UtcNow;
                 }));
 
                 if (!string.IsNullOrWhiteSpace(newPost.ImageFileName))
@@ -75,22 +76,34 @@ namespace Profitable.Services.Posts
             }
         }
 
-        public async Task<Result> DeletePostAsync(Guid guid)
+        public async Task<Result> DeletePostAsync(Guid guid, Guid requesterGuid)
         {
-            var post = await postsRepository
+            try
+            {
+                var post = await postsRepository
                  .GetAllAsNoTracking()
                  .FirstOrDefaultAsync(entity => entity.Guid == guid);
 
-            if (post == null)
-            {
-                return GlobalServicesConstants.EntityDoesNotExist("Post");
+                if (post == null)
+                {
+                    return GlobalServicesConstants.EntityDoesNotExist("Post");
+                }
+
+                if (post.AuthorId != requesterGuid)
+                {
+                    return GlobalServicesConstants.RequesterNotOwnerMesssage;
+                }
+
+                postsRepository.Delete(post);
+
+                await postsRepository.SaveChangesAsync();
+
+                return true;
             }
-
-            postsRepository.Delete(post);
-
-            await postsRepository.SaveChangesAsync();
-
-            return true;
+            catch (Exception e)
+            {
+                return e.Message;
+            }
         }
 
         public async Task<PostResponseModel> GetPostByGuidAsync(
@@ -207,23 +220,31 @@ namespace Profitable.Services.Posts
             }
         }
 
-        public async Task<Result> UpdatePostAsync(string postToUpdateGuid, UpdatePostRequestModel newPost)
+        public async Task<Result> UpdatePostAsync(Guid postGuid, UpdatePostRequestModel newPost, Guid requesterGuid)
         {
-            if (newPost.Content.Length > GlobalServicesConstants.PostMaxLength)
+            try
             {
-                throw new ArgumentException(
-                    $"Content must be no longer than {GlobalServicesConstants.PostMaxLength} characters.");
-            }
+                if (newPost.Content.Length > GlobalServicesConstants.PostMaxLength)
+                {
+                    throw new ArgumentException(
+                        $"Content must be no longer than {GlobalServicesConstants.PostMaxLength} characters.");
+                }
 
-            var postToUpdateGuidCasted = Guid.Parse(postToUpdateGuid);
+                var postToUpdate = await postsRepository
+                    .GetAll()
+                    .Where(post => !post.IsDeleted)
+                    .FirstOrDefaultAsync(post => post.Guid == postGuid);
 
-            var postToUpdate = await postsRepository
-                .GetAll()
-                .Where(post => !post.IsDeleted)
-                .FirstOrDefaultAsync(post => post.Guid == postToUpdateGuidCasted);
+                if (postToUpdate == null)
+                {
+                    return GlobalServicesConstants.EntityDoesNotExist("Post");
+                }
 
-            if (postToUpdate != null)
-            {
+                if (postToUpdate.AuthorId != requesterGuid)
+                {
+                    return GlobalServicesConstants.RequesterNotOwnerMesssage;
+                }
+
                 await imageService.DeleteUploadedImageAsync(ImageFor.Posts, postToUpdate.ImageURL);
 
                 string newFileName = "";
@@ -246,26 +267,24 @@ namespace Profitable.Services.Posts
 
                 return true;
             }
-            else
+            catch (Exception e)
             {
-                return GlobalServicesConstants.EntityDoesNotExist("Post");
+                return e.Message;
             }
         }
 
-        public async Task<int> ManagePostLikeAsync(ApplicationUser author, string postGuid)
+        public async Task<int> ManagePostLikeAsync(Guid authorId, Guid postGuid)
         {
-            var postToUpdateGuidCasted = Guid.Parse(postGuid);
-
             var postToUpdate = await postsRepository
                  .GetAllAsNoTracking()
                  .Where(post => !post.IsDeleted)
                  .Include(post => post.Likes)
-                 .FirstOrDefaultAsync(post => post.Guid == postToUpdateGuidCasted);
+                 .FirstOrDefaultAsync(post => post.Guid == postGuid);
 
             var postLikeExisted = await likesRepository
                 .GetAllAsNoTracking()
                 .FirstOrDefaultAsync(postLike =>
-                    postLike.PostId == postToUpdateGuidCasted && postLike.AuthorId == author.Id);
+                    postLike.PostId == postGuid && postLike.AuthorId == authorId);
 
             int likesToShow = postToUpdate.Likes.Count;
 
@@ -273,8 +292,8 @@ namespace Profitable.Services.Posts
             {
                 var likeToAdd = new Like()
                 {
-                    PostId = postToUpdateGuidCasted,
-                    AuthorId = author.Id,
+                    PostId = postGuid,
+                    AuthorId = authorId,
                 };
 
                 await likesRepository.AddAsync(likeToAdd);
